pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragrag"
dynamic = ["version"]
requires-python = ">=3.10,<3.15"
dependencies = [
    "qdrant-client>=1.12.0,<1.17",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.40.0,<5",
    "accelerate>=0.26.0",
    "Pillow>=10.0.0",
    "PyMuPDF>=1.24.0",
    "pytesseract>=0.3.10",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "qwen-vl-utils>=0.0.10",
    "python-magic>=0.4.27",
]

[project.scripts]
ragrag = "ragrag.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-timeout>=2.0.0",
    "PyYAML>=6.0",
    "nox>=2024.3",
    "pytest-cov>=4.0",
    "coverage[toml]>=7.0",
]

[tool.setuptools.dynamic]
version = {attr = "ragrag.__version__"}

[tool.setuptools.packages.find]
where = ["."]
include = ["ragrag*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 300

[tool.coverage.run]
branch = true
source = ["ragrag"]
omit = ["tests/*", "ragrag/embedding/colqwen_embedder.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ ==",
]

[tool.coverage.paths]
source = ["ragrag/", "*/site-packages/ragrag/"]
